Move gamePlay directory to /src/
