+++ src/gamePlay/boardGrid/boardScene.java
@@ -1,18 +1,14 @@
-package boardGrid;
+package gamePlay.boardGrid;
 
-import javafx.application.Application;
 import javafx.scene.Scene;
 import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
-import javafx.stage.StageStyle;
 
-import java.io.IOException;
 import java.util.ArrayList;
 
-public class boardLauncher extends Application {
-    @Override
-    public void start(Stage stage) throws IOException {
+public class boardScene {
 
+    public static Scene makeScene(Stage stage){
         int[] size = {8, 7};
         ArrayList<ArrayList<StackPane>> stackTable = new ArrayList<>();
         for (int i = 0; i < size[0]; i++){
@@ -29,14 +25,6 @@ public class boardLauncher extends Application {
         playerList.add("Marius");
 
         Scene scene = gamePlayUI.makeScene(stage, playerList, stackTable);
-        stage.setTitle("Game Board");
-
-
-        stage.setScene(scene);
-        stage.show();
-    }
-
-    public static void main(String[] args) {
-        launch();
+        return scene;
     }
 }
